+++ docs/changelog.md
@@ -9,7 +9,9 @@
 - *Sequencer* - Added `.scrollingText()` which allows to play scrolling text on the canvas for users
 - *Sequencer* - Added `.toJSON()` and `.fromJSON()` to Sequences to be able to be serialized and deserialized; only sequences with effects, sounds, and scrolling texts can be serialized
 - *Sequencer* - Added options to `.play()`, which may contain an object; currently supports `{ remote: true/false }` which will serialize the sequence (see above), and send it to each client for local playback, instead of the person running the sequence sending data to clients as it is being executed 
-- *Effects* - Improved how effects are replicated on linked tokens when `.persist()`'s `persistPrototypeToken` is enabled
+- *Sequencer* - Added database support for `_timestamps` metadata on effect files, which will trigger the `sequencerEffectTimestamp` hook when effects reach the point of the timestamps for that file 
+- *Sequencer* - Added support for flipbook-type effects
+- *Effects* - Vastly improved how effects are replicated on linked tokens when `.persist()`'s `persistPrototypeToken` is enabled
 
 ### Version 2.414
 - *Sequencer* - Included missing CSS file